requirements.txt
Flask>=2.0
pycryptodome>=3.15
python-pptx>=0.6.21
reportlab>=3.6.12
streamlit>=1.20
